add distinct login and register pages
